requirements.txt
numpy
torch
scikit-learn
matplotlib
pyarrow
pillow
mpi4py
petsc4py
fenics-dolfinx[all]